basic variable substitution
Instead of writing policies with (quantifier + marker), introduce variables in a toml file and write policies with those instead.
Does not yet support proper scoping--this commit is just about introducing variable names.
When compiling, naively substitutes (quantifier + marker) everywhere the variable appears, so it is functionally the same as previous commits.