lemmy community policy v1
This isn't the ideal version of the policy -- it repeats the ban check count for every delete check,
and deals with closing the parentheses in a brittle way in the conditional body.
The larger problem is that while compiling this way preserves correctness for this particular policy,
that wouldn't hold for other policies. Whether or not we run the ban_check code depends on the existence of delete_check nodes,
 which is not the policy writer's intention. It's fine for this policy, because the function call is any(), so if there are no
delete check nodes we'll fail. But if the function call switches to delete_checks.all(), then the whole policy is true if there are no delete checks, even if the ban_checks remains any().  Going to move to let bindings in the body to deal with this--people need
to specify where the scopes of their variables begin and end.